src/webapp/requirements.txt
azure-identity>=1.15.0
azure-search-documents>=11.6.0
azure-storage-blob>=12.19.0
openai>=1.12.0
quart>=0.19.0
python-dotenv>=1.0.0
hypercorn>=0.16.0
aiohttp>=3.9.0
PyPDF2>=3.0.0
